# ⚡️ Speed up `from_values`