--- a/lamindb/_can_validate.py
+++ b/lamindb/_can_validate.py
@@ -375,12 +375,18 @@ def _standardize(
         # here, we can safely import bionty
         from bionty._bionty import create_or_get_organism_record
 
-        organism_record = create_or_get_organism_record(
-            organism=organism, registry=registry
-        )
-        organism = (
-            organism_record.name if organism_record is not None else organism_record
-        )
+        if (
+            hasattr(registry, "_ontology_id_field")
+            and field == registry._ontology_id_field
+        ):
+            pass
+        else:
+            organism_record = create_or_get_organism_record(
+                organism=organism, registry=registry
+            )
+            organism = (
+                organism_record.name if organism_record is not None else organism_record
+            )
 
     # only perform synonym mapping if field is the name field
     if hasattr(registry, "_name_field") and field != registry._name_field: