--- a/pyproject.toml
+++ b/pyproject.toml
@@ -18,7 +18,7 @@ dependencies = [
     # Lamin PINNED packages
     "lnschema_core==0.74.5",
     "lamindb_setup==0.77.5",
-    "lamin_utils==0.13.4",
+    "lamin_utils==0.13.5",
     "lamin_cli==0.17.6",
     # others
     "rapidfuzz",