--- a/tests/core/test_feature_manager.py
+++ b/tests/core/test_feature_manager.py
@@ -18,6 +18,7 @@ def adata():
 # below the test for the main way of annotating with
 # features
 def test_features_add(adata):
+    ln.Feature.filter().delete()
     ln.ULabel(name="Experiment 1")
     artifact = ln.Artifact.from_anndata(adata, description="test")
     artifact.save()