--- a/tests/core/test_feature_manager.py
+++ b/tests/core/test_feature_manager.py
@@ -18,8 +18,11 @@ def adata():
 # below the test for the main way of annotating with
 # features
 def test_features_add(adata):
+    ln.FeatureSet.filter().delete()
+    ln.Artifact.filter().delete(permanent=True)
     ln.Feature.filter().delete()
-    ln.ULabel(name="Experiment 1")
+    ln.ULabel.filter().delete()
+
     artifact = ln.Artifact.from_anndata(adata, description="test")
     artifact.save()
     with pytest.raises(ValidationError) as error: