3.7s → 1.3s from public source (records not yet registered):
<img width="769" alt="Screenshot 2024-09-29 at 16 03 46" src="https://github.com/user-attachments/assets/f1a2f751-27aa-47da-88eb-1482809b8243">

1.1s → 0.1s from lamindb registry (records already registered):
<img width="717" alt="Screenshot 2024-09-29 at 16 07 21" src="https://github.com/user-attachments/assets/208bd8b3-a05d-44c8-890f-5a49c6d2d412">