@falexwolf 2024-09-26
Are you going to move it to `CanValidate` in this PR or another one?